// src/module/admin/admin.interfaces.ts
import { Document } from 'mongoose';

export interface IAdmin extends Document {
  // 姓名
  readonly username: string;
  // 邮箱
  readonly email: string;
  // 联系电话
  readonly phone: string;
  // 密码
  password: string;
  // 头像
  readonly avatar: string;
  // 注册时间
  readonly registerTime: Date;
  // 最后登录时间
  readonly lastLoginTime: Date;
  // 最后登录ip
  readonly lastLoginIp: string;

  accessToken: string;

  // 权限
  readonly role: number;

  readonly isDelete: boolean;
  readonly deleteTime: Date;
}